+++ python/ray/data/_internal/datasource/audio_datasource.py
@@ -1,10 +1,10 @@
 import io
 from typing import TYPE_CHECKING, Iterator, List, Union
 
+from ray.data._internal.datasource.file_based_datasource import FileBasedDatasource
 from ray.data._internal.delegating_block_builder import DelegatingBlockBuilder
 from ray.data._internal.util import _check_import
 from ray.data.block import Block
-from ray.data.datasource.file_based_datasource import FileBasedDatasource
 
 if TYPE_CHECKING:
     import pyarrow
