+++ python/ray/data/_internal/datasource/bigquery_datasink.py
@@ -12,11 +12,11 @@ if TYPE_CHECKING:
 
 import ray
 from ray.data._internal.datasource import bigquery_datasource
+from ray.data._internal.datasource.datasink import Datasink
 from ray.data._internal.execution.interfaces import TaskContext
 from ray.data._internal.remote_fn import cached_remote_fn
 from ray.data._internal.util import _check_import
 from ray.data.block import Block, BlockAccessor
-from ray.data.datasource.datasink import Datasink
 
 logger = logging.getLogger(__name__)
