+++ python/ray/data/_internal/datasource/delta_sharing_datasource.py
@@ -4,9 +4,9 @@ from typing import TYPE_CHECKING, List, Optional, Tuple
 
 import numpy as np
 
+from ray.data._internal.datasource.datasource import Datasource, ReadTask
 from ray.data._internal.util import _check_import
 from ray.data.block import BlockMetadata
-from ray.data.datasource.datasource import Datasource, ReadTask
 
 if TYPE_CHECKING:
     from ray.data.context import DataContext
