+++ python/ray/data/_internal/datasource/hudi_datasource.py
@@ -3,9 +3,9 @@ import os
 from enum import Enum
 from typing import TYPE_CHECKING, Dict, Iterator, List, Optional, Tuple
 
+from ray.data._internal.datasource.datasource import Datasource, ReadTask
 from ray.data._internal.util import _check_import
 from ray.data.block import BlockMetadata
-from ray.data.datasource.datasource import Datasource, ReadTask
 
 if TYPE_CHECKING:
     from ray.data.context import DataContext
