+++ python/ray/data/_internal/datasource/huggingface_datasource.py
@@ -1,11 +1,11 @@
 import sys
 from typing import TYPE_CHECKING, Iterable, List, Optional, Union
 
+from ray.data._internal.datasource import Datasource, ReadTask
 from ray.data._internal.tensor_extensions.arrow import pyarrow_table_from_pydict
 from ray.data._internal.util import _check_pyarrow_version
 from ray.data.block import Block, BlockAccessor, BlockMetadata
 from ray.data.dataset import Dataset
-from ray.data.datasource import Datasource, ReadTask
 
 if TYPE_CHECKING:
     import datasets
