+++ python/ray/data/_internal/datasource/kafka_datasource.py
@@ -30,11 +30,11 @@ import pyarrow as pa
 if TYPE_CHECKING:
     from kafka import KafkaConsumer, TopicPartition
 
+from ray.data._internal.datasource import Datasource, ReadTask
 from ray.data._internal.output_buffer import BlockOutputBuffer, OutputBlockSizeOption
 from ray.data._internal.util import _check_import
 from ray.data.block import Block, BlockMetadata
 from ray.data.context import DataContext
-from ray.data.datasource import Datasource, ReadTask
 
 logger = logging.getLogger(__name__)
