+++ python/ray/data/_internal/datasource/range_datasource.py
@@ -4,10 +4,10 @@ from typing import Iterable, List, Optional, Tuple
 
 import numpy as np
 
+from ray.data._internal.datasource import Datasource, ReadTask
 from ray.data._internal.util import _check_pyarrow_version
 from ray.data.block import Block, BlockAccessor, BlockMetadata
 from ray.data.context import DataContext
-from ray.data.datasource import Datasource, ReadTask
 
 
 class RangeDatasource(Datasource):
