+++ python/ray/data/_internal/datasource/torch_datasource.py
@@ -1,8 +1,8 @@
 from typing import TYPE_CHECKING, Optional
 
+from ray.data._internal.datasource.datasource import Datasource, ReadTask
 from ray.data._internal.delegating_block_builder import DelegatingBlockBuilder
 from ray.data.block import BlockMetadata
-from ray.data.datasource.datasource import Datasource, ReadTask
 
 if TYPE_CHECKING:
     import torch
