+++ python/ray/data/datasource/__init__.py
@@ -1,15 +1,7 @@
-from ray.data._internal.datasource.delta_sharing_datasource import (
-    DeltaSharingDatasource,
-)
-from ray.data._internal.datasource.mcap_datasource import (
-    MCAPDatasource,
-    TimeRange,
-)
-from ray.data._internal.datasource.sql_datasource import Connection
-from ray.data._internal.datasource.turbopuffer_datasink import (
-    TurbopufferDatasink,
-)
-from ray.data._internal.savemode import SaveMode
+# Lazy imports for symbols from _internal to avoid circular import:
+# ray.data._internal.datasource is loaded by ray.data.__init__, which then
+# loads submodules (e.g. audio_datasource) that import from this package.
+# If we import _internal here at top level, _internal is not fully loaded yet.
 from ray.data.datasource.datasink import (
     Datasink,
     DummyOutputDatasink,
@@ -47,6 +39,48 @@ from ray.data.datasource.partitioning import (
 # Note: HuggingFaceDatasource should NOT be imported here, because
 # we want to only import the Hugging Face datasets library when we use
 # ray.data.from_huggingface() or HuggingFaceDatasource() directly.
+
+# Lazy-loaded names from _internal (see __getattr__ below)
+_INTERNAL_NAMES = frozenset(
+    {
+        "Connection",
+        "DeltaSharingDatasource",
+        "MCAPDatasource",
+        "TimeRange",
+        "SaveMode",
+        "TurbopufferDatasink",
+    }
+)
+
+
+def __getattr__(name: str):
+    if name in _INTERNAL_NAMES:
+        if name == "TurbopufferDatasink":
+            from ray.data._internal.datasource.turbopuffer_datasink import (
+                TurbopufferDatasink,
+            )
+
+            return TurbopufferDatasink
+        if name == "SaveMode":
+            from ray.data._internal.savemode import SaveMode
+
+            return SaveMode
+        from ray.data._internal.datasource import (
+            Connection,
+            DeltaSharingDatasource,
+            MCAPDatasource,
+            TimeRange,
+        )
+
+        return {
+            "Connection": Connection,
+            "DeltaSharingDatasource": DeltaSharingDatasource,
+            "MCAPDatasource": MCAPDatasource,
+            "TimeRange": TimeRange,
+        }[name]
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+
+
 __all__ = [
     "BaseFileMetadataProvider",
     "Connection",
